multiple databases support, agency module

+++ deploy/local/windows/reinstall.ps1
@@ -1,7 +1,10 @@
-
+#!/usr/bin/env powershell
+param ($force)
 .\stop.ps1
 
-"docker volume prune -f" | cmd
-rm -r -fo volumes
+if($force) {
+    "docker volume prune -f" | cmd
+    rm -r -fo volumes
+}
 
 .\setup.ps1
